+++ tests/Unit/Framework/Config/ConfigInitTest.php
@@ -83,27 +83,25 @@ final class ConfigInitTest extends TestCase
 
     public function test_read_single_config(): void
     {
-        $gacelaJsonConfigCreator = $this->createStub(GacelaConfigFileFactoryInterface::class);
-        $gacelaJsonConfigCreator
-            ->method('createGacelaFileConfig')
-            ->willReturn((new GacelaConfigFile())
-                ->setConfigItems([
-                    'supported-type' => new GacelaConfigItem('supported-type'),
-                ]));
-
         $reader = $this->createStub(ConfigReaderInterface::class);
         $reader->method('canRead')->willReturn(true);
         $reader->method('read')->willReturn(['key' => 'value']);
 
-        $readers = [
-            'supported-type' => $reader,
-        ];
+        $gacelaJsonConfigCreator = $this->createStub(GacelaConfigFileFactoryInterface::class);
+        $gacelaJsonConfigCreator
+            ->method('createGacelaFileConfig')
+            ->willReturn((new GacelaConfigFile())
+                ->setConfigReaders([
+                    'supported-type' => $reader,
+                ])
+                ->setConfigItems([
+                    'supported-type' => new GacelaConfigItem('supported-type'),
+                ]));
 
         $configInit = new ConfigLoader(
             'application_root_dir',
             $gacelaJsonConfigCreator,
             $this->createMock(PathFinderInterface::class),
-            $readers
         );
 
         self::assertSame(['key' => 'value'], $configInit->loadAll());
@@ -111,15 +109,6 @@ final class ConfigInitTest extends TestCase
 
     public function test_read_multiple_config(): void
     {
-        $gacelaJsonConfigCreator = $this->createStub(GacelaConfigFileFactoryInterface::class);
-        $gacelaJsonConfigCreator
-            ->method('createGacelaFileConfig')
-            ->willReturn((new GacelaConfigFile())
-                ->setConfigItems([
-                    'supported-type1' => new GacelaConfigItem('supported-type1'),
-                    'supported-type2' => new GacelaConfigItem('supported-type2'),
-                ]));
-
         $reader1 = $this->createStub(ConfigReaderInterface::class);
         $reader1->method('canRead')->willReturn(true);
         $reader1->method('read')->willReturn(['key1' => 'value1']);
@@ -128,16 +117,23 @@ final class ConfigInitTest extends TestCase
         $reader2->method('canRead')->willReturn(true);
         $reader2->method('read')->willReturn(['key2' => 'value2']);
 
-        $readers = [
-            'supported-type1' => $reader1,
-            'supported-type2' => $reader2,
-        ];
+        $gacelaJsonConfigCreator = $this->createStub(GacelaConfigFileFactoryInterface::class);
+        $gacelaJsonConfigCreator
+            ->method('createGacelaFileConfig')
+            ->willReturn((new GacelaConfigFile())
+                ->setConfigReaders([
+                    'supported-type1' => $reader1,
+                    'supported-type2' => $reader2,
+                ])
+                ->setConfigItems([
+                    'supported-type1' => new GacelaConfigItem('supported-type1'),
+                    'supported-type2' => new GacelaConfigItem('supported-type2'),
+                ]));
 
         $configInit = new ConfigLoader(
             'application_root_dir',
             $gacelaJsonConfigCreator,
             $this->createMock(PathFinderInterface::class),
-            $readers
         );
 
         self::assertSame([